CDAP-19705 refactored Column View code
