+++ app/cdap/components/ColumnView/__test__/ColumnView.test.tsx
@@ -17,7 +17,69 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import ColumnView from 'components/ColumnView';
 import React from 'react';
-import { mockColumnData, mockDataQuality } from 'components/ColumnView/mock/mockDataForColumnView';
+
+export const mockColumnData = [
+  {
+    name: 'body_0',
+    label: 'body_0',
+    type: ['Int'],
+  },
+  {
+    name: 'body_1',
+    label: 'body_1',
+    type: ['String'],
+  },
+  {
+    name: 'body_2',
+    label: 'body_2',
+    type: ['String'],
+  },
+  {
+    name: 'body_3',
+    label: 'body_3',
+    type: ['String'],
+  },
+];
+
+export const mockDataQuality = {
+  body_0: {
+    general: {
+      'non-null': 100,
+    },
+  },
+  body_1: {
+    general: {
+      'non-null': 66.66667,
+      null: 33.333336,
+    },
+  },
+  body_2: {
+    general: {
+      'non-null': 83.33333,
+      null: 16.666668,
+    },
+    types: {
+      Integer: 33.333336,
+      Text: 33.333336,
+    },
+  },
+  body_3: {
+    general: {
+      'non-null': 50,
+      null: 50,
+    },
+    types: {
+      Text: 16.666668,
+    },
+  },
+};
+
+export const mockResult = [
+  { label: 'body_0', value: 0 },
+  { label: 'body_1', value: 33.333336 },
+  { label: 'body_2', value: 16.666668 },
+  { label: 'body_3', value: 50 },
+];
 
 describe('It Should test Column View Component', () => {
   it('Should render the Column View Component and the column view panel , select Column list component to be in the document', () => {
