--- app/cdap/components/ColumnView/constants.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-import T from 'i18n-react';
-
-export const COLUMNS = T.translate('features.WranglerNewUI.ColumnViewPanel.columns');
-export const HEADING_TEXT = T.translate('features.WranglerNewUI.ColumnViewPanel.columnView');
-export const NULL_VALUES = T.translate('features.WranglerNewUI.ColumnViewPanel.nullValues');
